+++ crates/ryo3-serde/src/constants.rs
@@ -1,0 +1,2 @@
+pub(crate) type Depth = u8;
+pub(crate) const MAX_DEPTH: Depth = 255;
