+++ crates/ryo3-serde/src/py_serialize.rs
@@ -3,7 +3,7 @@
 //! Based on a combination of `orjson`, `pythonize` and `rtoml`.
 
 use pyo3::prelude::*;
-use serde::ser::{Error as SerError, Serialize, SerializeMap, Serializer};
+use serde::ser::{Error as SerError, Serialize, Serializer};
 
 use crate::any_repr::any_repr;
 use crate::errors::pyerr2sererr;
@@ -13,26 +13,12 @@ use crate::pytypes::{
 };
 #[cfg(feature = "ry")]
 use crate::rytypes;
-// #[cfg(feature = "ryo3-ulid")]
-// use crate::rytypes::ry_ulid;
-// #[cfg(feature = "ryo3-url")]
-// use crate::rytypes::ry_url;
-// #[cfg(feature = "ryo3-uuid")]
-// use crate::rytypes::ry_uuid;
-// #[cfg(feature = "ryo3-jiff")]
-// use crate::rytypes::{
-//     ry_date, ry_datetime, ry_signed_duration, ry_span, ry_time, ry_timestamp, ry_timezone, ry_zoned,
-// };
-// #[cfg(feature = "ryo3-http")]
-// use crate::rytypes::{ry_headers, ry_http_status};
+use crate::safe_impl::{DataclassSerializer, SerializePyMapping, SerializePySequence};
 use crate::type_cache::{PyObType, PyTypeCache};
-use pyo3::sync::GILOnceCell;
-use pyo3::types::{PyAnyMethods, PyDict, PyMapping, PySequence, PyString};
-use pyo3::{Bound, intern};
-use serde::ser::SerializeSeq;
+use crate::{Depth, MAX_DEPTH};
+use pyo3::Bound;
+use pyo3::types::{PyAnyMethods, PyDict, PyMapping, PySequence};
 
-type Depth = u8;
-const MAX_DEPTH: Depth = 255;
 pub struct SerializePyAny<'py> {
     pub(crate) obj: &'py Bound<'py, PyAny>,
     pub(crate) depth: Depth,
@@ -82,6 +68,7 @@ impl<'py> SerializePyAny<'py> {
         }
     }
 }
+
 fn dataclass_fields<'a, 'py>(obj: &'a Bound<'py, PyAny>) -> Option<Bound<'py, PyDict>>
 where
     'py: 'a, // keep lifetimes compatible
@@ -173,173 +160,6 @@ impl Serialize for SerializePyAny<'_> {
             self.with_obj(&r).serialize(serializer)
         } else {
             serde_err!("{} is not json-serializable", any_repr(self.obj))
-        }
-    }
-}
-
-// ===========================================================================
-// PySequence ~ PySequence ~ PySequence ~ PySequence ~ PySequence ~ PySequence
-// ===========================================================================
-struct SerializePySequence<'a, 'py> {
-    seq: &'a Bound<'py, PySequence>,
-    depth: Depth,
-    default: Option<&'py Bound<'py, PyAny>>,
-}
-
-impl<'a, 'py> SerializePySequence<'a, 'py> {
-    fn new_with_depth(
-        seq: &'a Bound<'py, PySequence>,
-        default: Option<&'py Bound<'py, PyAny>>,
-        depth: Depth,
-    ) -> Self {
-        Self {
-            seq,
-            depth,
-            default,
-        }
-    }
-}
-
-impl Serialize for SerializePySequence<'_, '_> {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let len = self.seq.len().map_err(pyerr2sererr)?;
-        let mut seq = serializer.serialize_seq(Some(len))?;
-        for i in 0..len {
-            let item = self.seq.get_item(i).map_err(pyerr2sererr)?;
-            let item_ser = SerializePyAny::new_with_depth(&item, self.default, self.depth + 1);
-            seq.serialize_element(&item_ser)?;
-        }
-        seq.end()
-    }
-}
-// ===========================================================================
-// PyMapping ~ PyMapping ~ PyMapping ~ PyMapping ~ PyMapping ~ PyMapping
-// ===========================================================================
-
-struct SerializePyMapping<'a, 'py> {
-    mapping: &'a Bound<'py, PyMapping>,
-    depth: Depth,
-    default: Option<&'py Bound<'py, PyAny>>,
-}
-
-impl<'a, 'py> SerializePyMapping<'a, 'py> {
-    fn new_with_depth(
-        mapping: &'a Bound<'py, PyMapping>,
-        default: Option<&'py Bound<'py, PyAny>>,
-        depth: Depth,
-    ) -> Self {
-        Self {
-            mapping,
-            depth,
-            default,
-        }
-    }
-}
-
-impl Serialize for SerializePyMapping<'_, '_> {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let len = self.mapping.len().ok();
-        if let Some(len) = len {
-            let mut m = serializer.serialize_map(Some(len))?;
-            let keys = self.mapping.keys().map_err(pyerr2sererr)?;
-            for k in keys {
-                let k = crate::pytypes::mapping_key(&k)?;
-                let val = self.mapping.get_item(k).map_err(pyerr2sererr)?;
-                let v = SerializePyAny::new_with_depth(&val, self.default, self.depth + 1);
-                m.serialize_entry(k, &v).map_err(pyerr2sererr)?;
-            }
-            m.end()
-        } else {
-            Err(S::Error::custom(
-                "SerializePyMapping: Length of mapping is not known.",
-            ))
-        }
-    }
-}
-
-// ===========================================================================
-// DATACLASSES
-// ===========================================================================
-
-struct DataclassSerializer<'a, 'py> {
-    obj: &'a Bound<'py, PyAny>,
-    default: Option<&'py Bound<'py, PyAny>>,
-    fields: Bound<'py, PyDict>,
-    depth: Depth,
-}
-
-impl<'a, 'py> DataclassSerializer<'a, 'py> {
-    fn new(
-        obj: &'a Bound<'py, PyAny>,
-        default: Option<&'py Bound<'py, PyAny>>,
-        depth: Depth,
-        fields: Bound<'py, PyDict>,
-    ) -> Self {
-        Self {
-            obj,
-            default,
-            fields,
-            depth,
-        }
-    }
-}
-// as done in pydantic-core: https://github.com/pydantic/pydantic-core/blob/5f0b5a8b26691b7a1e3de07cb409b21bb174929c/src/serializers/shared.rs#L591
-static DC_FIELD_MARKER: GILOnceCell<PyObject> = GILOnceCell::new();
-/// needed to match the logic from dataclasses.fields `tuple(f for f in fields.values() if f._field_type is _FIELD)`
-fn get_field_marker(py: Python<'_>) -> PyResult<&Bound<'_, PyAny>> {
-    DC_FIELD_MARKER.import(py, "dataclasses", "_FIELD")
-}
-
-impl Serialize for DataclassSerializer<'_, '_> {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        if self.depth == MAX_DEPTH {
-            return Err(SerError::custom("recursion"));
-        }
-
-        // check for __dict__
-        if let Ok(dunder_dict) = self.obj.getattr("__dict__") {
-            if let Ok(dict) = dunder_dict.downcast_into::<PyDict>() {
-                // serialize the __dict__ as a dict
-                SerializePyAny::new_with_depth(&dict, self.default, self.depth + 1)
-                    .serialize(serializer)
-            } else {
-                serde_err!("__dict__ is not a dict")
-            }
-        } else {
-            let py = self.obj.py();
-            let field_marker = get_field_marker(py).map_err(pyerr2sererr)?;
-            let mut map = serializer.serialize_map(None)?;
-            for (field_name, field) in self.fields.iter() {
-                // check if the field is a dataclass field
-                let field_type = field
-                    .getattr(intern!(py, "_field_type"))
-                    .map_err(pyerr2sererr)?;
-                if field_type.is(field_marker) {
-                    // this is a dataclass field
-                    let field_name_py_str = field_name
-                        .downcast_into::<PyString>()
-                        .map_err(pyerr2sererr)?;
-                    let value = self.obj.getattr(&field_name_py_str).map_err(pyerr2sererr)?;
-                    let field_ser =
-                        SerializePyAny::new_with_depth(&value, self.default, self.depth + 1);
-
-                    // actual string
-                    let s = field_name_py_str
-                        .to_str()
-                        .map_err(|_| SerError::custom("field name is not a valid UTF-8 string"))?;
-                    map.serialize_entry(s, &field_ser)?;
-                }
-            }
-            map.end()
         }
     }
 }
