+++ crates/ryo3-serde/src/ser/mod.rs
@@ -1,0 +1,3 @@
+//! Serialize python objects
+//!
+//! Currently pre-reorg
